requirements.txt
alembic==1.0.10
amqp==2.5.0
apturl==0.5.2
asn1crypto==0.24.0
billiard==3.6.0.0
Brlapi==0.6.6
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
command-not-found==0.3
configparser==3.7.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.1
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
fdfs-client-py==1.2.6
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.12
MarkupSafe==1.1.1
mutagen==1.42.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
py3Fdfs==2.2.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.7.2
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyRFC3339==1.0
python-alipay-sdk==1.10.1
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-editor==1.0.4
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
qiniu==7.2.4
redis==3.2.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
uWSGI==2.0.18
vine==1.3.0
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==0.15.4
Whoosh==2.7.4
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
